requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.28.1
psycopg==3.2.3
redis==5.2.0
celery==5.4.0
gunicorn==23.0.0
loguru==0.7.2
pycryptodomex==3.18.0